+++ apps/web/playwright/apps/conferencing/types.ts
@@ -1,0 +1,6 @@
+export type TApp = {
+  slug: string;
+  type: string;
+  organizerInputPlaceholder?: string;
+  label: string;
+};
