+++ apps/web/playwright/fixtures/apps.ts
@@ -1,6 +1,6 @@
 import { expect, type Page } from "@playwright/test";
 
-import type { TApp } from "../apps/conferencing/conferencingApps.e2e";
+import type { TApp } from "../apps/conferencing/types";
 import {
   bookTimeSlot,
   gotoBookingPage,
