+++ apps/web/playwright/fixtures/eventTypes.ts
@@ -1,6 +1,6 @@
 import { expect, type Page } from "@playwright/test";
 
-import { localize } from "../lib/testUtils";
+import { localize } from "../lib/localize";
 
 export function createEventTypeFixture(page: Page) {
   return {
