+++ apps/web/playwright/lib/testUtils.ts
@@ -14,7 +14,6 @@ import { BookingStatus, SchedulingType } from "@calcom/prisma/enums";
 
 import type { createEmailsFixture } from "../fixtures/emails";
 import type { Fixtures } from "./fixtures";
-import { loadJSON } from "./loadJSON";
 
 type Request = IncomingMessage & { body?: unknown };
 type RequestHandlerOptions = { req: Request; res: ServerResponse };
@@ -176,16 +175,6 @@ export async function expectSlotNotAllowedToBook(page: Page) {
     return response.url().includes("/slots/isAvailable");
   });
   await expect(page.locator("[data-testid=slot-not-allowed-to-book]")).toBeVisible();
-}
-
-// Provide an standalone localize utility not managed by next-i18n
-export async function localize(locale: string) {
-  const localeModule = `../../public/static/locales/${locale}/common.json`;
-  const localeMap = loadJSON(localeModule);
-  return (message: string) => {
-    if (message in localeMap) return localeMap[message];
-    throw "No locale found for the given entry message";
-  };
 }
 
 export const createNewEventType = async (page: Page, args: { eventTitle: string }) => {
