+++ apps/web/playwright/login.e2e.ts
@@ -2,7 +2,7 @@ import { expect } from "@playwright/test";
 
 import { login } from "./fixtures/users";
 import { test } from "./lib/fixtures";
-import { localize } from "./lib/testUtils";
+import { localize } from "./lib/localize";
 
 test.describe.configure({ mode: "parallel" });
