+++ packages/features/bookings/lib/handleCancelBooking.ts
@@ -6,7 +6,6 @@ import { DailyLocationType } from "@calcom/app-store/locations";
 import dayjs from "@calcom/dayjs";
 import { sendCancelledEmailsAndSMS } from "@calcom/emails";
 import { getCalEventResponses } from "@calcom/features/bookings/lib/getCalEventResponses";
-import { workflowSelect } from "@calcom/features/ee/workflows/lib/getAllWorkflows";
 import { sendCancelledReminders } from "@calcom/features/ee/workflows/lib/reminders/reminderScheduler";
 import type { GetSubscriberOptions } from "@calcom/features/webhooks/lib/getWebhooks";
 import getWebhooks from "@calcom/features/webhooks/lib/getWebhooks";
@@ -25,132 +24,21 @@ import { safeStringify } from "@calcom/lib/safeStringify";
 import { getTranslation } from "@calcom/lib/server/i18n";
 import { WorkflowRepository } from "@calcom/lib/server/repository/workflow";
 import { getTimeFormatStringFromUserTimeFormat } from "@calcom/lib/timeFormat";
-import prisma, { bookingMinimalSelect } from "@calcom/prisma";
+import prisma from "@calcom/prisma";
 import type { WebhookTriggerEvents } from "@calcom/prisma/enums";
 import { BookingStatus } from "@calcom/prisma/enums";
-import { credentialForCalendarServiceSelect } from "@calcom/prisma/selects/credential";
 import { bookingMetadataSchema, EventTypeMetaDataSchema, bookingCancelInput } from "@calcom/prisma/zod-utils";
 import type { EventTypeMetadata } from "@calcom/prisma/zod-utils";
 import { getAllWorkflowsFromEventType } from "@calcom/trpc/server/routers/viewer/workflows/util";
 import type { CalendarEvent } from "@calcom/types/Calendar";
 
 import { getAllCredentials } from "./getAllCredentialsForUsersOnEvent/getAllCredentials";
+import { getBookingToDelete } from "./getBookingToDelete";
+import type { CustomRequest, AppRouterRequest } from "./handleCancelBooking/types";
 import { handleInternalNote } from "./handleInternalNote";
 import cancelAttendeeSeat from "./handleSeats/cancel/cancelAttendeeSeat";
 
 const log = logger.getSubLogger({ prefix: ["handleCancelBooking"] });
-
-async function getBookingToDelete(id: number | undefined, uid: string | undefined) {
-  return await prisma.booking.findUniqueOrThrow({
-    where: {
-      id,
-      uid,
-    },
-    select: {
-      ...bookingMinimalSelect,
-      recurringEventId: true,
-      userId: true,
-      user: {
-        select: {
-          id: true,
-          username: true,
-          credentials: { select: credentialForCalendarServiceSelect }, // Not leaking at the moment, be careful with
-          email: true,
-          timeZone: true,
-          timeFormat: true,
-          name: true,
-          destinationCalendar: true,
-        },
-      },
-      location: true,
-      references: {
-        select: {
-          uid: true,
-          type: true,
-          externalCalendarId: true,
-          credentialId: true,
-          thirdPartyRecurringEventId: true,
-          delegationCredentialId: true,
-        },
-      },
-      payment: true,
-      paid: true,
-      eventType: {
-        select: {
-          slug: true,
-          owner: {
-            select: {
-              id: true,
-              hideBranding: true,
-            },
-          },
-          team: {
-            select: {
-              id: true,
-              name: true,
-              parentId: true,
-            },
-          },
-          parentId: true,
-          parent: {
-            select: {
-              teamId: true,
-            },
-          },
-          userId: true,
-          recurringEvent: true,
-          title: true,
-          eventName: true,
-          description: true,
-          requiresConfirmation: true,
-          price: true,
-          currency: true,
-          length: true,
-          seatsPerTimeSlot: true,
-          bookingFields: true,
-          seatsShowAttendees: true,
-          metadata: true,
-          schedulingType: true,
-          hosts: {
-            select: {
-              user: true,
-            },
-          },
-          workflows: {
-            select: {
-              workflow: {
-                select: workflowSelect,
-              },
-            },
-          },
-        },
-      },
-      uid: true,
-      id: true,
-      eventTypeId: true,
-      destinationCalendar: true,
-      smsReminderNumber: true,
-      workflowReminders: true,
-      seatsReferences: true,
-      responses: true,
-      iCalUID: true,
-      iCalSequence: true,
-    },
-  });
-}
-
-export type BookingToDelete = Awaited<ReturnType<typeof getBookingToDelete>>;
-
-export type AppRouterRequest = { appDirRequestBody: unknown };
-export type CustomRequest = (NextApiRequest | AppRouterRequest) & {
-  userId?: number;
-  bookingToDelete?: BookingToDelete;
-  platformClientId?: string;
-  platformRescheduleUrl?: string;
-  platformCancelUrl?: string;
-  platformBookingUrl?: string;
-  arePlatformEmailsEnabled?: boolean;
-};
 
 export type HandleCancelBookingResponse = {
   success: boolean;
